--- packages/react-ui/.storybook/preview.ts
@@ -1,17 +1,0 @@
-import type { Preview } from "@storybook/react";
-import { decorators } from "./decorator";
-
-const preview: Preview = {
-  decorators,
-  parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/,
-      },
-    },
-  },
-};
-
-export default preview;
